+++ api/pong/modules/Auth/serializers.py
@@ -1,0 +1,23 @@
+from django.contrib.auth import authenticate
+from rest_framework import serializers
+
+
+class LoginSerializer(serializers.Serializer):
+    username = serializers.CharField(required=True)
+    password = serializers.CharField(required=True, write_only=True)
+
+    def validate(self, attrs):
+        username = attrs.get("username")
+        password = attrs.get("password")
+
+        if username and password:
+            user = authenticate(username=username, password=password)
+            if user:
+                if user.is_active:
+                    attrs["user"] = user
+                    return attrs
+                raise serializers.ValidationError("User account is disabled.")
+            raise serializers.ValidationError(
+                "Unable to log in with provided credentials."
+            )
+        raise serializers.ValidationError('Must include "username" and "password".')
